Fixes #1068 - Pass the correct local time zone for calendar entry creation
Actual fix suggested by @exoticDFT.  Previously "America/New_York" was passed as the time zone to create calendar entries from the "Add to Calendar" button, however this was generating the calendar entry then with the wrong local time (e.g., it was set an hour off in a non-Eastern time zone).  Instead, pass the user's timezone.